Making the version type long. Also updating to commons-lang 2.5

+++ trunk/greenbridge-junit/src/main/java/com/googlecode/greenbridge/annotation/ScenarioRef.java
@@ -9,7 +9,7 @@ public interface ScenarioRef {
 
     public String[] narrative();
     public String name();
-    public int version();
+    public long version();
     public String linkUrl();
     public String linkName();
     public Class<? extends StoryRef> getStoryRef();
